+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "hw_js",
+  "version": "1.0.0",
+  "description": "HW_JS",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest",
+    "start": "webpack serve --open",
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/GalkinMM/HW_JS.git"
+  },
+  "author": "GalkinMM",
+  "license": "ISC",
+  "devDependencies": {
+    "css-loader": "^6.7.1",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0"
+  },
+  "bugs": {
+    "url": "https://github.com/GalkinMM/HW_JS/issues"
+  },
+  "homepage": "https://github.com/GalkinMM/HW_JS#readme"
+}
