+++ server/public/getBasket.json
@@ -1,1 +1,1 @@
-[{"id_product":1,"count":1},{"id_product":2,"count":1},{"id_product":3,"count":1},{"id_product":4,"count":1}]
+[{"id_product":1,"count":7},{"id_product":4,"count":8},{"id_product":3,"count":11},{"id_product":2,"count":4}]
