+++ src/components/customSearch/index.js
@@ -1,0 +1,10 @@
+// @ts-nocheck
+export default Vue.component("custom_search", {
+    template: `
+        <input
+            type="text"
+            class="search__input"
+            placeholder="Поиск"
+            @input="$emit('input', $event.target.value)">
+    `
+});
