+++ src/constances.js
@@ -1,0 +1,4 @@
+export const BASE_URL = 'http://localhost:8000/';
+export const getGoodsIiems = `${BASE_URL}catalogData.json`;
+export const GOODS = `${BASE_URL}goods`;
+export const BASKETGOODS = `${BASE_URL}basket`;
