+++ src/index.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <link rel="stylesheet" href="style/style.css">
+
+    <script src="https://cdn.jsdelivr.net/npm/vue@2.6.14/dist/vue.js"></script>
+    <script src="script/main.js"></script>
+
+    <title>eShop</title>
+</head>
+<body>
+    <div id="root">
+        <header>
+            <div class="container">
+                <h1 class="title">eShop</h1>
+
+                <div class="search">
+                    <custom_search v-model="searchValue"></custom_search>
+                </div>
+
+                <button class="cartButton header__cartButton"
+                        @click="basketWindowVisibleChange">
+                        Корзина
+                </button>
+            </div>
+        </header>
+    
+        <main>
+            <div class="container">
+                <div class="goodsList" v-if="filteredList.length">
+                    <vc_gooditems
+                        v-for="item in filteredList"
+                        :item="item">
+                    </vc_gooditems>
+                </div>
+                <div v-else>Товаров нет</div>
+            </div>
+            <vc_basket
+                v-if="basketWindowVisible" 
+                @close="basketWindowVisibleChange">
+            </vc_basket>
+        </main>
+    </div>
+</body>
+</html>
