+++ src/style/scss/_main.scss
@@ -1,0 +1,32 @@
+/* Стили main */
+main {
+    .container {
+        padding: 32px;
+    }
+
+    .goodsList {
+        display: grid;
+        grid-auto-flow: row;
+        grid-template-columns: auto max-content;
+        gap: 16px;
+        justify-content: center;
+    }
+
+    .goodsItem {
+        border: 1px solid black;
+        border-radius: 8px;
+        max-width: 256px;
+        padding: 16px;
+        text-align: right;
+
+        img {
+            max-width: 100%;
+        }
+
+        img,
+        h3,
+        p{
+            padding-bottom: 4px;
+        }
+    }
+}
